% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnorm0.R
\name{dnorm0}
\alias{dnorm0}
\title{Density for Zero-Inflated Gaussian Distribution}
\usage{
dnorm0(y, alpha, mu, sd1, log = F)
}
\arguments{
\item{alpha}{the inflated zero component.}

\item{mu}{the mean of the Gaussian part.}

\item{log}{calculates the density on the log scale or not. The default is False.}

\item{x}{the vector of observations.}

\item{sigma}{the standard deviation of the Gaussian part.}
}
\value{
the vector of calculated densities for vector \code{x}; of the same length as \code{x}.
}
\description{
\code{dnorm0} calculates the density for zero-inflated Gaussian distribution.
}
\examples{
dnorm0(x=c(1,3,5,7),alpha=0.3,mu=5,sigma=1,log=F)

}
\keyword{Gaussian}
\keyword{Zero-Inflated}
