% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dZITGaussian.R
\name{dZITGaussian}
\alias{dZITGaussian}
\title{Density for Zero-Inflated Truncated Gaussian Distribution}
\usage{
dZITGaussian(x, alpha, mu, sigma, log = F)
}
\arguments{
\item{x}{the vector of observations.}

\item{alpha}{the inflated zero component.}

\item{mu}{the mean of the truncated Gaussian part.}

\item{sigma}{the standard deviation of the truncated Gaussian part.}

\item{log}{calculates the density on the log scale or not. The default is False.}
}
\value{
the vector of calculated densities for vector \code{x}; of the same length as \code{x}.
}
\description{
\code{dZITGaussian} calculates the density for zero-inflated truncated Gaussian distribution.
}
\examples{
dZITGaussian(x=c(1,3,5,7),alpha=0.3,mu=5,sigma=1,log=F)

}
\keyword{Gaussian}
\keyword{Truncated}
\keyword{Zero-Inflated}
