% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian.R
\name{Gaussian}
\alias{Gaussian}
\title{auxillary function for fitting Gaussian distribution}
\usage{
Gaussian(y, weights = NULL)
}
\arguments{
\item{y}{the vector of observations for fitting zero-inflated truncated Gaussian.}

\item{weights}{an optional vector of ‘prior weights’ to be used in the fitting process. Should be NULL or a numeric vector.}
}
\value{
a list of two elements:

\item{coefficients}{mean.}

\item{dispersion}{standard deviation.}
}
\description{
auxillary function for fitting Gaussian distribution
}
\keyword{internal}
