% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_ZITG.R
\name{HMM_ZITG}
\alias{HMM_ZITG}
\title{The HMM-based Sleep/Wake Identification Algorithm Assuming Zero-Inflated Truncated Gaussian}
\usage{
HMM_ZITG(Yob, paramInit, tranInit, piInit, ntime = NULL, maxIter = 20,
  reltol = 10^-2)
}
\arguments{
\item{Yob}{the vector of observations from wearable devices.}

\item{paramInit}{the list of initial HMM parameters: list(c(\eqn{\alpha},\eqn{\mu_1},\eqn{\sigma_1}),c(\eqn{\mu_2},\eqn{\sigma_2})).}

\item{tranInit}{the initial 2 by 2 transition matrix.}

\item{piInit}{the initial probabilities of length 2: \eqn{\pi_1},\eqn{\pi_2}.}

\item{ntime}{the vector specifying the start of each independent time series of observations. If not specified, the observations are considered as one single time series.}

\item{maxIter}{the maximal number of iteration in the fitting procedure. The default is 20.}

\item{reltol}{the relative tolerance of log likelihood in the fitting procedure. The default is 10^-2.}
}
\value{
a list. Assume that \code{s} steps in the fitting procedure have been performed:

\item{alpha11}{the estimated zero component \eqn{\alpha_1} for the sleep state (state 1); a vector of length \code{s}.}

\item{mu1}{the estimated mean \eqn{\mu_1} for the sleep state (state 1); a vector of length \code{s}.}

\item{sd1}{the estimated standard deviation \eqn{\sigma_1} for the sleep state (state 1); a vector of length \code{s}.}

\item{mu2}{the estimated mean \eqn{\mu_2} for the wake state (state 2); a vector of length \code{s}.}

\item{sd2}{the estimated standard deviation \eqn{\sigma_2} for the wake state (state 2); a vector of length \code{s}.}

\item{loglik}{the estimated log likelihood; a vector of length \code{s}.}

\item{transit}{the estimated transition matrix; a list of length \code{s}.}

\item{init}{the estimated initial probabilities; a list of length \code{s}.}

\item{iterMax}{logical; if TRUE, the maximal number of iterations have been reached.}
}
\description{
In the sleep/wake identification algorithm, the HMM assumes two states: sleep (state 1)
and wake (state 2). State 1 assumes zero-inflated truncated Gaussian, while state 2
assumes Gaussian.
}
\details{
It is noted that the estimation of truncated Gaussian using
the function \code{truncreg.fit} in the package \code{truncreg} is sometimes unstable.
Thus, we recommend using \code{HMM_ZIG}.
}
\examples{
data(ob)
ob <- ob[1:840] #first day observation

piI <- c(0.9,0.1)  #state 1 and 2
paramI <- list(c(0.4,3,1),  #state 1
               c(6,1))      #state 2
tranI <- matrix(c(0.9,0.1,0.2,0.8),nrow=2,ncol=2,byrow=T)

re <- HMM_ZITG(Yob=ob[1:840],paramInit=paramI,tranInit=tranI,piInit=piI,ntime=c(1),maxIter=30,reltol=10^-6)

}
\keyword{Gaussian}
\keyword{Truncated}
\keyword{Zero-Inflated}
