% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.data.R
\name{read.data}
\alias{read.data}
\title{The function to prepare data before implementing the HMM-based Sleep/Wake Identification Algorithm directly without GGIR pre-processing, i.e. prepare summary count data. The example format comes from NHANES.}
\usage{
read.data(filename, directory = "", ...)
}
\arguments{
\item{filename}{The name of the file containing summary count data to be reformatted.}

\item{directory}{The directory in which the file is stored.}
}
\value{
a csv file named "nightsummary.csv" by default that contains sleep information for all individuals.
}
\description{
In the sleep/wake identification algorithm, the HMM assumes two states: sleep (state 1)
and wake (state 2). As the activity levels follow different distributions under different
states, HMM can differentiate the two states accordingly.
}
\examples{
#df_ex <- read.data("example.csv", "C:/myfolder", sep=",", header=T, stringsAsFactors = F)
}
\keyword{sleep}
\keyword{variables}
