% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm0.R
\name{rnorm0}
\alias{rnorm0}
\title{Random Generation for the Zero-Inflated Gaussian Distribution (Approximation)}
\usage{
rnorm0(n, alpha, mu, sd1, positive)
}
\arguments{
\item{n}{the number of observations.}

\item{alpha}{the inflated zero component.}

\item{mu}{the mean of the truncated Gaussian part.}

\item{positive}{whether only positive values are allowed. The default is TRUE.}

\item{sigma}{the standard deviation of the truncated Gaussian part.}
}
\value{
the vector of generated observations of length \code{n}.
}
\description{
\code{dnorm0} random generation for the zero-inflated Gaussian distribution (approximation).
}
\examples{
rnorm0(n=100,alpha=0.3,mu=5,sigma=1,positive=TRUE)

}
\keyword{Gaussian}
\keyword{Zero-Inflated}
