% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rZITGaussian.R
\name{rZITGaussian}
\alias{rZITGaussian}
\title{Random Generation for the Zero-Inflated Truncated Gaussian Distribution}
\usage{
rZITGaussian(n, alpha, mu, sigma)
}
\arguments{
\item{n}{the number of observations.}

\item{alpha}{the inflated zero component.}

\item{mu}{the mean of the truncated Gaussian part.}

\item{sigma}{the standard deviation of the truncated Gaussian part.}
}
\value{
the vector of generated observations of length \code{n}.
}
\description{
\code{rZITGaussian} random generation for the zero-inflated truncated Gaussian distribution.
}
\examples{
rZITGaussian(n=100,alpha=0.3,mu=5,sigma=1)

}
\keyword{Gaussian}
\keyword{Zero-Inflated}
\keyword{truncated}
