% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.part3.R
\name{sleep.part3}
\alias{sleep.part3}
\title{The HMM-based Sleep/Wake Identification Algorithm to identify sustained inactive periods}
\usage{
sleep.part3(
  metadatadir = c(),
  f0 = 1,
  f1 = 0,
  lastdayend = NULL,
  nonwearhour = 2,
  ninvalid = 16,
  startearly = NULL,
  endearly = NULL,
  ignorenonwear = TRUE,
  overwrite = FALSE,
  do.pdf = TRUE,
  do.parallel = TRUE,
  desiredtz = "",
  ...
)
}
\arguments{
\item{metadatadir}{Directory that holds a folder ’meta’, which is created after sleep.prep.}

\item{f0}{File index to start with (default = 1, and usually working with the default is fine). Index refers to the file names sorted in the increasing order.}

\item{f1}{File index to finish with (defaults to the number of files available, and setting to 0 is fine).}

\item{lastdayend}{A variable to suggest the approximate end time on the last day. Participants often take off the watch early on the last day but the data may continue to be collected and this period may be failed to be detected as non-wear periods. Therefore, lastdayend can help prevent this problem by setting an approximate end time. If lastdayend = 9, then presumably the participant takes off the watch around 9am. The default is NULL.}

\item{nonwearhour}{Remove the detected nonwear intervals that are shorter than L hours. The default value of \code{L} is 2 (hours).}

\item{ninvalid}{Remove the days that have < \code{h} hours valid data. The default is \code{h}=16 hours.}

\item{startearly}{a safeguard to prevent evening nap/take-off period from being identified as sleep onset time. For example, sometimes the shower watch take-off is not identified and removed as a non-wear period, and is then recognized as the sleep onset. If the safeguard startearly = 20, then the marker for sleep start shall be after 20:00/8pm. The default is NULL.}

\item{endearly}{a safeguard to prevent wake-up at night from being identified as morning wake-up time. For example, if the safeguard endearly = 5, then the marker for sleep end shall be after 5am. The default is NULL.}

\item{ignorenonwear}{If ignorenonwear=TRUE then ignore detected monitor non-wear periods before identifying sustained inactivity (default = TRUE).}

\item{overwrite}{Whether to overwrite data that were generated by this function previously. If overwrite = FALSE then it will not overwrite but skip the previously generated files (default = FALSE).}

\item{do.pdf}{Whether to generate a pdf for sleep quality control visualization. Turning this off could speed up the process.}

\item{do.parallel}{Whether to use multi-core processing. Parallel processing only works if at least 4 CPU cores are available (default = TRUE).}

\item{desiredtz}{desired timezone.}
}
\value{
a RData file containing information on sustained inactive periods. No direct output.
}
\description{
In the sleep/wake identification algorithm, the HMM assumes two states: sleep (state 1)
and wake (state 2). As the activity levels follow different distributions under different
states, HMM can differentiate the two states accordingly.
}
\details{
This function helps identify sustained inactive periods as input for the downstream functions.
}
\examples{

#metadir = "C:/myfolder/meta" # the meta folder containing the results from previous steps
#sleep.part3(metadatadir=metadir, overwrite=TRUE)

}
\keyword{inactive}
\keyword{periods}
