% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.part5.R
\name{sleep.part5}
\alias{sleep.part5}
\title{The HMM-based Sleep/Wake Identification Algorithm to output sleep variables}
\usage{
sleep.part5(
  metadatadir = c(),
  f0 = 1,
  f1 = 0,
  datadir = c(),
  daysleeplength = 0.5,
  nonwearhour = 2,
  overwrite = FALSE,
  do.parallel = TRUE
)
}
\arguments{
\item{metadatadir}{Directory that holds a folder ’meta’.}

\item{f0}{File index to start with (default = 1, and usually working with the default is fine). Index refers to the file names sorted in the increasing order.}

\item{f1}{File index to finish with (defaults to the number of files available, and setting to 0 is fine).}

\item{datadir}{Directory where the accelerometer files are stored or list of accelerometer filenames and directories.}

\item{daysleeplength}{the threshold for calculating daytime sleep. The default value is 0.5 in the unit of hour (= 30 minutes), meaning that daytime sleep episodes that are longer than 30 minutes will be considered.}

\item{nonwearhour}{Remove the detected nonwear intervals that are shorter than L hours. The default value of \code{L} is 2 (hours).}

\item{overwrite}{Whether to overwrite data that were generated by this function previously. If overwrite = FALSE then it will not overwrite but skip the previously generated files (default = FALSE).}

\item{do.parallel}{Whether to use multi-core processing. Parallel processing only works if at least 4 CPU cores are available (default = TRUE).}
}
\value{
a dataset on sleep variables.

\item{sleeponset}{the evening sleep onset timing for the night sleep, expressed in the unit of hour. For example, value = 21 means 21:00/9pm.}

\item{wakeup}{the morning wakeup timing for the night sleep, expressed in the unit of hour. For example, value = 8 means 08:00/8am.}

\item{SptDuration}{the sleep duration, calculated as the time between sleep onset and wakeup, expressed in the unit of hour. For example, value = 10 means 10 hours of nighttime sleep.}

\item{guider_onset}{HMM crude guider of the sleep onset timing, expressed in the unit of hour. For example, value = 21 means 21:00/9pm.}

\item{guider_wakeup}{HMM crude guider of the wakeup timing, expressed in the unit of hour. For example, value = 8 means 08:00/8am.}

\item{guider_SptDuration}{The crude estimate of sleep duration by HMM crude guiders of sleep onset and wakeup.}

\item{sleeponset_ts}{the sleep onset timing, expressed in the format hh:mm:ss.}

\item{wakeup_ts}{the wakeup timing, expressed in the format hh:mm:ss.}

\item{guider_onset_ts}{HMM guider of the sleep onset timing, expressed in the format hh:mm:ss.}

\item{guider_wakeup_ts}{HMM guider of the wakeup timing, expressed in the format hh:mm:ss.}

\item{sleep_true}{calculated as the amount of true sleep time in the defined sleep duration period.}

\item{sleep_efficiency}{calculated as the amount of true sleep time divided by the night sleep duration.}

\item{waso}{wake after sleep onset, calculated as the amount of wake time in the defined sleep duration period.}

\item{sleep_daytime_Nmin_more}{daytime sleep time, only counting the period that is longer than N minutes. The default N = 30 minutes.}

\item{sleep_day_episode_Nmin_more}{the number of daytime sleep episodes, only counting the period that is longer than N minutes. The default N = 30 minutes.}

\item{sleep_total}{total sleep time – 24 hr assessment of sleep.}

\item{L5_midpoint}{the midpoint of L5, which is the least 5 active hour period of the 24 hour day. For example, value = 2 means that the midpoint of L5 is 2 hours/2am.}

\item{M10_midpoint}{the midpoint of M10, which is the most active 10 hour period of the 24 hour day. For example, value = 16 means that the midpoint of M10 is 16 hours/4pm. Limited by each day data, NA will be reported as -999.}
}
\description{
In the sleep/wake identification algorithm, the HMM assumes two states: sleep (state 1)
and wake (state 2). As the activity levels follow different distributions under different
states, HMM can differentiate the two states accordingly.
}
\examples{
#metadir = "C:/myfolder/meta" # the meta folder containing the results from previous steps
#sleep.part5(metadatadir=metadir, overwrite=TRUE)

}
\keyword{sleep}
\keyword{variables}
