% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.part6.R
\name{sleep.part6}
\alias{sleep.part6}
\title{The HMM-based Sleep/Wake Identification Algorithm to summarize the sleep information of all individuals and output in a csv file.}
\usage{
sleep.part6(metadatadir = c(), outputdir = c())
}
\arguments{
\item{metadatadir}{Directory that holds a folder ’meta’.}

\item{outputdir}{Directory that the output csv file will be saved into.}
}
\value{
a csv file named allsummary.csv that contains sleep information for all individuals.
}
\description{
In the sleep/wake identification algorithm, the HMM assumes two states: sleep (state 1)
and wake (state 2). As the activity levels follow different distributions under different
states, HMM can differentiate the two states accordingly.
}
\examples{
#metadir = "C:/myfolder/meta" # the meta folder containing the results from previous steps
#sleep.part6(metadatadir=metadir)

}
\keyword{sleep}
\keyword{variables}
