% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.prep.R
\name{sleep.prep}
\alias{sleep.prep}
\title{The HMM-based Sleep/Wake Identification Algorithm to provide sleep information: preparation step}
\usage{
sleep.prep(
  datadir = c(),
  outputdir = c(),
  f0 = 1,
  f1 = 0,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{datadir}{Directory where the accelerometer files are stored.}

\item{outputdir}{Directory where the output needs to be stored. Folders are created in this directory to store outputs}

\item{f0}{File index to start with (default = 1). Index refers to the filenames sorted in increasing order}

\item{f1}{File index to finish with (defaults to number of files available)}

\item{overwrite}{Overwrite previously generated milestone data by this function for this particular dataset. If FALSE then it will skip the previously processed files (default = FALSE).}
}
\value{
No output will be generated. It is a pre-processing step of accelerometer data in preparation for downstream analysis sleep.part3 through sleep.part6.
}
\description{
It uses the shell function in GGIR to read in accelerometer/actigraph data and pre-process using R. The pre-processed data can be further used for sleep/wake identification.
}
\examples{
#datadir = "C:/myfolder/datadir"
#outputdir = "C:/myfolder/outputdir"
#sleep.prep(datadir=datadir,
#outputdir=outputdir,
#f0=1,f1=0, overwrite=TRUE)

}
\keyword{sleep}
\keyword{variables}
