% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.summary.R
\name{sleep.summary}
\alias{sleep.summary}
\title{The function to implement the HMM-based Sleep/Wake Identification Algorithm directly without GGIR pre-processing, i.e. using summary count data. The example format comes from NHANES.}
\usage{
sleep.summary(
  df,
  interval = 60,
  minhour = 16,
  maxtry = 20,
  filename = "",
  outputdir = ""
)
}
\arguments{
\item{df}{The directory that holds a folder ’meta’.}

\item{interval}{The summary count data interval in seconds, i.e. summary count data every 5 seconds, 30 seconds, 60 seconds (interval=60). The default is interval=60 (60 seconds or 1 minute).}

\item{minhour}{The minimal number of hours available to be considered for night analysis; otherwise sleep parameters will not be calculated for the night. The default is minhour = 16.}

\item{maxtry}{The maximal number of trials that the program will run to fit HMM. Randomness is involved in initiation and there can be convergence issues due to data quality. The default is maxtry = 20.}

\item{filename}{The name of the file to store results. The default is "nightsummary.csv".}

\item{outputdir}{The directory that the output csv file will be saved into.}
}
\value{
a csv file named "nightsummary.csv" by default that contains sleep information for all individuals.
}
\description{
In the sleep/wake identification algorithm, the HMM assumes two states: sleep (state 1)
and wake (state 2). As the activity levels follow different distributions under different
states, HMM can differentiate the two states accordingly.
}
\examples{
#df_ex <- read.data("example.csv", "C:/myfolder", sep=",", header=T, stringsAsFactors = F)
#sleep.summary(df_ex,interval = 60, minhour = 16, maxtry = 20, filename = "example.csv", outputdir = "C:/outdir")

}
\keyword{sleep}
\keyword{variables}
