% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIGaussian.R
\name{ZIGaussian}
\alias{ZIGaussian}
\title{Fitting Zero-Inflated Gaussian Distribution}
\usage{
ZIGaussian(y, weights = NULL, maxstep = 100, reltol = 10^-6)
}
\arguments{
\item{y}{the vector of observations for fitting zero-inflated Gaussian.}

\item{weights}{an optional vector of ‘prior weights’ to be used in the fitting process. Should be NULL or a numeric vector.}

\item{maxstep}{the maximal number of steps for fitting the Gaussian part and the zero part. The default is 100.}

\item{reltol}{the relative tolerance of log likelihood in the fitting procedure. The default is 10^-6.}
}
\value{
a list of two elements.

\item{count}{a list of two elements for the Gaussian part: mean and sigma.}

\item{zero}{the estimated proportion of zeros for the zero inflation part.}
}
\description{
\code{ZIGaussian} is used to fit zero-inflated Gaussian distribution.
}
\examples{
data(ob)
ZIGaussian(ob,weights=NULL,maxstep=100,reltol=10^-6)

}
\keyword{Gaussian}
\keyword{Zero-Inflated}
