% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZITGaussian.R
\name{ZITGaussian}
\alias{ZITGaussian}
\title{Fitting Zero-Inflated Truncated Gaussian Distribution}
\usage{
ZITGaussian(y, weights = NULL, maxstep = 100, reltol = 10^-6)
}
\arguments{
\item{y}{the vector of observations for fitting zero-inflated truncated Gaussian.}

\item{weights}{an optional vector of ‘prior weights’ to be used in the fitting process. Should be NULL or a numeric vector.}

\item{maxstep}{the maximal number of steps for fitting the truncated Gaussian part and the zero part. The default is 100.}

\item{reltol}{the relative tolerance of log likelihood in the fitting procedure. The default is 10^-6.}
}
\value{
a list of two elements.

\item{count}{a list of two elements for the truncated Gaussian part: mean and sigma.}

\item{zero}{the estimated proportion of zeros for the zero inflation part.}
}
\description{
\code{ZITGaussian} is used to fit zero-inflated truncated Gaussian distribution.
}
\examples{
ZITGaussian(y,weights=NULL,maxstep=100,reltol=10^-6)

}
\keyword{Gaussian}
\keyword{Truncated}
\keyword{Zero-Inflated}
